consolidated the apps, now 3 apps but still the 6 pages that can be accessed through different functions in the apps
