--- finalProj/createaccount/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import createaccountPageView
-
-urlpatterns = [
-    path('', createaccountPageView, name='createaccount')
-]
