--- finalProj/creategroup/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import creategroupPageView
-
-urlpatterns = [
-    path('', creategroupPageView, name='creategroup')
-]
