--- finalProj/creategroup/views.py
@@ -1,5 +1,0 @@
-from django.shortcuts import render
-from django.http import HttpResponse
-
-def creategroupPageView(request) :
-    return HttpResponse('This is where you create a new group')
