+++ finalProj/finalProj/settings.py
@@ -40,9 +40,6 @@ INSTALLED_APPS = [
     'loginpage.apps.LoginpageConfig',
     'viewgroups.apps.ViewgroupsConfig',
     'viewclasses.apps.ViewclassesConfig',
-    'createaccount.apps.CreateaccountConfig',
-    'creategroup.apps.CreategroupConfig',
-    'addclass.apps.AddclassConfig',
 ]
 
 MIDDLEWARE = [
