+++ finalProj/loginpage/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
-from .views import indexPageView
+from .views import indexPageView, createaccountPageView
 
 urlpatterns = [
-    path('', indexPageView, name='index')
+    path('', indexPageView, name='index'),
+    path('createaccount/', createaccountPageView, name='createaccount')
 ]
