+++ finalProj/loginpage/views.py
@@ -3,3 +3,6 @@ from django.http import HttpResponse
 
 def indexPageView(request) :
     return HttpResponse('Login Page')
+
+def createaccountPageView(request) :
+    return HttpResponse('This is where you create an account')
