+++ finalProj/viewclasses/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
-from .views import viewclassesPageView
+from .views import viewclassesPageView, addclassPageView
 
 urlpatterns = [
-    path('', viewclassesPageView, name='viewclasses')
+    path('', viewclassesPageView, name='viewclasses'),
+    path('addclass/', addclassPageView, name='addclass'),
 ]
