+++ finalProj/viewgroups/views.py
@@ -3,3 +3,6 @@ from django.http import HttpResponse
 
 def viewgroupsPageView(request) :
     return HttpResponse('View Groups Here')
+
+def creategroupPageView(request) :
+    return HttpResponse('This is where you create a new group')
